pyproject.toml
[project]
name = "analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "ipykernel>=6.29.5",
    "jinja2>=3.1.4",
    "numpy>=2.1.3",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pip>=24.3.1",
    "psutil>=6.1.0",
    "pyyaml>=6.0.2",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
]
